Refactored the name of the ProjectList component.
