+++ components/sections/project-list/ProjectList.types.ts
@@ -1,0 +1,11 @@
+import { IProject } from "../../../models/IProject";
+
+/**
+ * @interface ProjectListProps
+ * @summary Project list props
+ * @author J. Trpka
+ * @property { Project[] } projects
+ */
+export interface ProjectListProps {
+  projects: IProject[]
+}
